refactor(padi): house PRT3 Preview open/validate under previewTab/
Feature-subsystem directory (code-police): target validation + open/close
writers live next to each other, distinct from the file-byte preview.ts leaf.
rightPanelStateEqual moves next to the chrome schema both writers share.
