+++ packages/padi/src/terminals.ts
@@ -15,15 +15,10 @@
  * state-reads + lifecycle from this file as a single module.
  */
 
-import { knownPorts, type TerminalId } from "@kolu/terminal-vocab/schema";
-import { ORPCError } from "@orpc/server";
+import type { TerminalId } from "@kolu/terminal-vocab/schema";
 import { notifyDirty } from "./publisher.ts";
 import { type SessionSnapshot, saveSession } from "./session/session.ts";
-import {
-  getTerminal,
-  listTerminals,
-  terminalEntries,
-} from "./terminal-registry.ts";
+import { getTerminal, terminalEntries } from "./terminal-registry.ts";
 import {
   beginSleepLocal,
   releaseSleptLocalPty,
@@ -41,11 +36,9 @@ import { updateClientMetadata } from "./terminalEndpoint/metadata.ts";
 // to preserve the metadata→local order the TDZ note above depends on.
 import { resolveTerminalEndpoint } from "./terminalEndpoint/resolve.ts";
 import {
-  DEFAULT_RIGHT_PANEL_PER_TERMINAL,
-  type PreviewLocation,
   type RightPanelPerTerminalState,
+  rightPanelStateEqual,
 } from "./chromeVocab.ts";
-import { assertPreviewTarget } from "./previewTarget.ts";
 import {
   composeTerminalMetadata,
   type CreateTerminalInput,
@@ -271,98 +264,6 @@ export function setRightPanelState(
   updateClientMetadata(entry, id, (m) => {
     m.rightPanel = next;
   });
-}
-
-/** Host-wide scanned listening ports — union across every active terminal's
- *  known sample. One fact ("what listens on this host"), derived next to the
- *  registry that owns the samples — not recomputed in the UI. */
-export function collectHostScannedPorts(): Set<number> {
-  const ports = new Set<number>();
-  for (const t of listTerminals()) {
-    const entry = getTerminal(t.id);
-    if (!entry || entry.meta.state !== "active") continue;
-    for (const p of knownPorts(entry.snapshot.ports)) {
-      ports.add(p.port);
-    }
-  }
-  return ports;
-}
-
-/** Open or navigate the Preview tab. `doorPorts` is the live-forward remote
- *  set for this host when the caller can see it (kolu-server composition);
- *  padi's own serve path passes an empty set — doors live on the app
- *  surface, and chip-driven previews are already in the scan set.
- *
- *  Validates via {@link assertPreviewTarget}, writes `preview`, switches
- *  `activeTab` to `"preview"`, uncollapses. A repeat call NAVIGATES. */
-export function previewOpen(
-  id: TerminalId,
-  input: { port: number; path: string },
-  allow: {
-    scannedPorts: ReadonlySet<number>;
-    doorPorts: ReadonlySet<number>;
-  } = { scannedPorts: collectHostScannedPorts(), doorPorts: new Set() },
-): PreviewLocation {
-  const entry = getTerminal(id);
-  if (!entry) {
-    throw new ORPCError("NOT_FOUND", {
-      message: `terminal ${id} not found`,
-    });
-  }
-  const cur = entry.meta.rightPanel;
-  const location = assertPreviewTarget({
-    port: input.port,
-    path: input.path,
-    scannedPorts: allow.scannedPorts,
-    doorPorts: allow.doorPorts,
-    currentPort: cur?.preview?.port ?? null,
-  });
-  const next: RightPanelPerTerminalState = {
-    ...(cur ?? DEFAULT_RIGHT_PANEL_PER_TERMINAL),
-    collapsed: false,
-    activeTab: "preview",
-    preview: location,
-  };
-  if (cur && rightPanelStateEqual(cur, next)) return location;
-  updateClientMetadata(entry, id, (m) => {
-    m.rightPanel = next;
-  });
-  return location;
-}
-
-/** Clear the Preview location. Leaves `activeTab` on `"preview"` so the empty
- *  frame / door-closed chrome still shows; the client may switch tabs after. */
-export function previewClose(id: TerminalId): void {
-  const entry = getTerminal(id);
-  if (!entry) {
-    throw new ORPCError("NOT_FOUND", {
-      message: `terminal ${id} not found`,
-    });
-  }
-  const cur = entry.meta.rightPanel;
-  if (!cur || cur.preview === null) return;
-  const next: RightPanelPerTerminalState = { ...cur, preview: null };
-  updateClientMetadata(entry, id, (m) => {
-    m.rightPanel = next;
-  });
-}
-
-function rightPanelStateEqual(
-  a: RightPanelPerTerminalState,
-  b: RightPanelPerTerminalState,
-): boolean {
-  if (a.collapsed !== b.collapsed) return false;
-  if (a.activeTab !== b.activeTab || a.codeMode !== b.codeMode) return false;
-  if ((a.preview?.port ?? null) !== (b.preview?.port ?? null)) return false;
-  if ((a.preview?.path ?? null) !== (b.preview?.path ?? null)) return false;
-  const am = a.selectedFileByMode;
-  const bm = b.selectedFileByMode;
-  if (am === bm) return true;
-  if (!am || !bm) return false;
-  if (am.local !== bm.local) return false;
-  if (am.branch !== bm.branch) return false;
-  if (am.browse !== bm.browse) return false;
-  return true;
 }
 
 // Active terminal ID — client-reported, used only for session snapshots.
